Adding csv for hack-a-thon collection.
First "by_subject" shall be used. After upcoming update "by_file_name" can be used to when processes are matched by file name.
To also get the onshore wind timeseries the collection.json, metadata for onshore wind and structure file have to be adjusted or wait for `additional_parameter` function to be refined.